refactor: improve type safety and reduce UI duplication
- use-tenant-payments.ts: Replace loose 'readonly unknown[]' with proper QueryKey type from @tanstack/react-query for better type safety and editor support
- manage/tenants/page.tsx: Replace three manually-rendered Card blocks with OwnerPaymentSummary component to centralize presentation logic
- Add OwnerPaymentSummary import and remove duplicate late fees/unpaid invoices/tenant count UI
- Fix page prerendering issues with proper dynamic exports and server/client component separation
- All quality gates passing: linting, typecheck, unit tests
