+++ apps/frontend/src/app/(protected)/manage/analytics/leases/page.tsx
@@ -1,3 +1,5 @@
+export const dynamic = 'force-dynamic'
+
 import { Badge } from '#components/ui/badge'
 import {
 	Card,
@@ -36,7 +38,7 @@ export default async function LeaseAnalyticsPage() {
 	return (
 		<div className="@container/main flex min-h-screen w-full flex-col">
 			<section
-				className="border-b bg-background p-6 border-(--color-fill-tertiary)"
+				className="border-b bg-background p-6 border-fill-tertiary"
 			>
 				<div className="mx-auto flex max-w-400 flex-col gap-6 px-4 lg:px-6">
 					<div className="flex flex-col gap-2">
