+++ apps/frontend/src/app/(protected)/manage/maintenance/page.tsx
@@ -17,21 +17,23 @@ export const metadata: Metadata = {
 export default async function MaintenancePage() {
 	// Server-side auth - NO client flash, instant 307 redirect
 	const { user } = await requireSession()
-	
+
 const logger = createLogger({ component: 'MaintenancePage', userId: user.id })
 
 	// Server Component: Fetch data on server during RSC render
 	let requests: MaintenanceRequestResponse['data'] = []
-	
+
 	try {
 		// Production pattern: Server Component with explicit token
 		const result: MaintenanceRequestResponse = await serverFetch('/api/v1/maintenance')
-		requests = result?.data ?? []
+		requests = Array.isArray(result?.data) ? result.data : []
 	} catch (err) {
 		// Log server-side; avoid throwing to prevent resetting the RSC tree
 		logger.warn('Failed to fetch maintenance requests for MaintenancePage', {
 			error: err instanceof Error ? err.message : String(err)
 		})
+		// Ensure requests is always an array
+		requests = []
 	}
 
 	return (
