+++ apps/frontend/src/app/(protected)/tenant/documents/page.tsx
@@ -1,4 +1,3 @@
-/**
 'use client'
 
 /**
