+++ apps/frontend/src/app/(protected)/tenant/page.tsx
@@ -136,7 +136,7 @@ export default function TenantDashboardPage() {
 					description="Upcoming rent payment"
 					className="dashboard-widget group hover:shadow-lg transition-all duration-300 hover:scale-[1.02] bg-linear-to-br from-green-50/50 to-card dark:from-green-950/20 dark:to-card border-2 hover:border-green-200 dark:hover:border-green-900"
 					footer={
-						<Link href="/tenant/payments">
+						<Link href="/tenant/payments/new">
 							<Button className="w-full">Pay Now</Button>
 						</Link>
 					}
@@ -193,7 +193,7 @@ export default function TenantDashboardPage() {
 				className="dashboard-section dashboard-widget border-2"
 			>
 				<div className="grid gap-5 sm:grid-cols-2 lg:grid-cols-4">
-					<Link href="/tenant/payments" className="group">
+					<Link href="/tenant/payments/new" className="group">
 						<Button
 							variant="outline"
 							className="w-full h-auto flex-col gap-3 py-8 border-2 hover:border-primary/40 hover:bg-accent/50 transition-all duration-300 hover:scale-105 hover:shadow-md"
